REFACTOR folder structure; clean up code
