+++ app/config/config.go
@@ -1,0 +1,1 @@
+package config
