+++ app/controller/story.controller.go
@@ -9,8 +9,8 @@ import (
 	"time"
 
 	"github.com/tolubydesign/angular-story-backend/app/models"
-	"github.com/tolubydesign/angular-story-backend/app/queries"
-	"github.com/tolubydesign/angular-story-backend/app/utils"
+	"github.com/tolubydesign/angular-story-backend/app/query"
+	"github.com/tolubydesign/angular-story-backend/app/database"
 	"github.com/tolubydesign/angular-story-backend/pkg/response"
 
 	"github.com/gofiber/fiber/v2"
@@ -21,7 +21,7 @@ import (
 func AllStoriesHandlerRequest(ctx *fiber.Ctx, db *sql.DB) error {
 	stories, err := queries.GetAllStories(db)
 
-	redisErr := utils.ConsoleActionToRedisDatabase("Attempting to get all stories")
+	redisErr := database.ConsoleActionToRedisDatabase("Attempting to get all stories")
 	if redisErr != nil {
 		return fiber.NewError(fiber.StatusInternalServerError, redisErr.Error())
 	}
@@ -57,7 +57,7 @@ func AllStoriesHandlerRequest(ctx *fiber.Ctx, db *sql.DB) error {
 		panic(err)
 	}
 
-	redisErr = utils.ConsoleActionToRedisDatabase("Request to get all stories successful")
+	redisErr = database.ConsoleActionToRedisDatabase("Request to get all stories successful")
 	if redisErr != nil {
 		return fiber.NewError(fiber.StatusInternalServerError, redisErr.Error())
 	}
